app/BackEnd/requirements.txt
Pillow
boto3
pandas 
PyYAML
fastapi[all]
uvicorn
